+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 1.3.1
+
+* Fix applying settings of link labels
+
+## 1.3.0
+
+* Add link data labels displaying
+
 ## 1.2.7
 
 * Fix scale settings text.